pyproject.toml
[tool.poetry]
name = "qa-guru-mobile-1"
version = "0.1.0"
description = ""
authors = ["Murat Kubekov <murat.kubekov@maximumtest.ru>"]
readme = "README.md"
packages = [{include = "qa_guru_mobile_1"}]

[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0-beta.16", allow-prereleases = true}
Appium-Python-Client = "^2.6.1"
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
